Cmor::All: Adapt to rails 6. Cmor::UserArea: Set minimum active_model_otp version to 2.1.1 as they added backup codes support that required changes to our code. Cmor::UserArea: Add one_time_password_options option as passthrough for active_model_otp options.

+++ cmor_blog/initialize_dummy.sh
@@ -7,11 +7,22 @@ rm -rf spec/dummy
 
 # Generate new dummy app
 DISABLE_MIGRATE=true rake dummy:app
+
+if [ ! -d "spec/dummy/config" ]; then exit 1; fi
+
+# Cleanup
 rm spec/dummy/.ruby-version
+rm spec/dummy/Gemfile
+
+cd spec/dummy
+
+# Use correct Gemfile
+sed -i "s|../Gemfile|../../../Gemfile|g" config/boot.rb
 
 # Satisfy prerequisites
-cd spec/dummy
 RAILS_ENV=development rails g model User email
+sed -i '17irequire "sassc-rails"' config/application.rb
+sed -i '17irequire "webpacker"' config/application.rb
 sed -i '17irequire "bootstrap4-kaminari-views"' config/application.rb
