+++ cmor_carousels_backend/initialize_dummy.sh
@@ -10,10 +10,17 @@ rm -rf spec/dummy
 
 # Generate new dummy app
 DISABLE_MIGRATE=true rake dummy:app
-rm spec/dummy/.ruby-version
 
-# Satisfy prerequisites
+if [ ! -d "spec/dummy/config" ]; then exit 1; fi
+
+# Cleanup
+rm spec/dummy/.ruby-version
+rm spec/dummy/Gemfile
+
 cd spec/dummy
+
+# Use correct Gemfile
+sed -i "s|../Gemfile|../../../Gemfile|g" config/boot.rb
 
 # responders for rao-service_controller
 ## Always require rspec and factory_bot_rails in dummy app
@@ -33,6 +40,9 @@ echo "end" >> config/initializers/route_translator.rb
 
 # Add ActiveStorage
 rails active_storage:install
+
+# Install simple form
+rails generate simple_form:install --bootstrap
 
 # Install cmor core backend gem
 rails generate administrador:install
