+++ cmor_suite/initialize_dummy.sh
@@ -6,12 +6,14 @@ INSTALL_NAME=${GEM_NAME//cmor_/cmor\:}
 rm -rf spec/dummy
 
 # Generate new dummy app
-DISABLE_MIGRATE=true bundle exec rake dummy:app
+DISABLE_MIGRATE=true rake dummy:app
 
+if [ ! -d "spec/dummy/config" ]; then exit 1; fi
+
+# Cleanup
 rm spec/dummy/.ruby-version
 rm spec/dummy/Gemfile
 
-# Satisfy prerequisites
 cd spec/dummy
 
 # Use correct Gemfile
@@ -22,7 +24,8 @@ rails generate simple_form:install --bootstrap
 
 # I don't know why we need this all of a sudden
 sed -i '17i\require "sassc-rails"' config/application.rb
-sed -i '18i\require "turbolinks"' config/application.rb
+sed -i '17i\require "turbolinks"' config/application.rb
+sed -i '17i\require "webpacker"' config/application.rb
 
 ## Always require rspec and factory_bot_rails in dummy app
 sed -i '17i\require "rspec-rails"' config/application.rb
